moved abel.vmi.find_image_center_by_slice() to abel.center. Uodated examples, and unit tests. tests.test_tools_vmi.py moved to tests.test_tools_center.py
