+++ abel/tests/test_tools_center.py
@@ -9,11 +9,11 @@ import numpy as np
 from numpy.testing import assert_allclose
 
 from abel.tools.analytical import sample_image
-from abel.tools.vmi import find_image_center_by_slice
+from abel.tools.center import find_image_center_by_slice
 
 from scipy.ndimage.interpolation import shift
 
-def test_vmi_find_image_center():
+def test_center_find_image_center():
 
     # BASEX sample image, Gaussians at 10, 15, 20, 70,85, 100, 145, 150, 155
     # image width, height n = 361
@@ -50,4 +50,4 @@ def test_vmi_find_image_center():
 
 
 if __name__ == "__main__":
-  test_vmi_find_image_center()
+  test_center_find_image_center()
